Restructure detestable to be test framework agnostic + support serialization for xunit
This commit splits the core running lnd declaration logic into a dependency free "Detestable" library.
Then it moves the xunit specific code to Detestable.Xunit. This means we
should be able to just add an adapater if we want to support other
testing frameworks like Nunit or MSBuild (no promises)